Added the .exe and bitaxe png
I will be adding the proper description on here as soon s I get my ducks in a row I want to make sure credit goes to the original author. This is just a GUI for the script, you still need python installed. It runs through freq. and core vol. and ramps up until a desired temp is met. It put out a .csv so you can see what optimal settings for the Bitaxe Gamma you have, they all different, a little ;>